fix: solve package connection problem
