+++ FileCache/Package.swift
@@ -4,15 +4,15 @@
 import PackageDescription
 
 let package = Package(
-    name: "FileCachePackage",
+    name: "FileCache",
     platforms: [
-        .iOS(.v13)
+        .iOS(.v16)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "FileCachePackage",
-            targets: ["FileCachePackage"])
+            name: "FileCache",
+            targets: ["FileCache"])
     ],
     dependencies: [
         .package(url: "https://github.com/CocoaLumberjack/CocoaLumberjack.git", from: "3.8.5")
@@ -21,12 +21,12 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "FileCachePackage",
+            name: "FileCache",
             dependencies: [
                 .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack")
             ]),
         .testTarget(
-            name: "FileCachePackageTests",
-            dependencies: ["FileCachePackage"])
+            name: "FileCacheTests",
+            dependencies: ["FileCache"])
     ]
 )
