+++ FileCache/README.md
@@ -1,3 +1,3 @@
-# FileCachePackage
+# FileCache
 
 A description of this package.
